WIP: Conv2D and GetCount

+++ include/fhenom/common.h
@@ -5,7 +5,7 @@
 #include <vector>
 
 namespace fhenom {
-using shape_t              = std::vector<size_t>;
+using shape_t              = std::vector<unsigned>;
 using Ctxt                 = lbcrypto::Ciphertext<lbcrypto::DCRTPoly>;
 using Ptxt                 = lbcrypto::Plaintext;
 using PrecomputedRotations = std::shared_ptr<std::vector<lbcrypto::DCRTPoly>>;
